+++ src/fleet/azext_fleet/vendored_sdks/models.py
@@ -4,4 +4,4 @@
 # Licensed under the MIT License. See License.txt in the project root for
 # license information.
 # --------------------------------------------------------------------------
-from .v2023_08_15_preview.models import *
+from .v2023_10_15.models import *
