+++ src/fleet/azext_fleet/vendored_sdks/v2023_10_15/_container_service_fleet_mgmt_client.py
@@ -7,13 +7,13 @@
 # --------------------------------------------------------------------------
 
 from copy import deepcopy
-from typing import Any, Awaitable, TYPE_CHECKING
+from typing import Any, TYPE_CHECKING
 
-from azure.core.rest import AsyncHttpResponse, HttpRequest
-from azure.mgmt.core import AsyncARMPipelineClient
+from azure.core.rest import HttpRequest, HttpResponse
+from azure.mgmt.core import ARMPipelineClient
 
-from .. import models as _models
-from ..._serialization import Deserializer, Serializer
+from . import models as _models
+from .._serialization import Deserializer, Serializer
 from ._configuration import ContainerServiceFleetMgmtClientConfiguration
 from .operations import (
     FleetMembersOperations,
@@ -25,35 +25,33 @@ from .operations import (
 
 if TYPE_CHECKING:
     # pylint: disable=unused-import,ungrouped-imports
-    from azure.core.credentials_async import AsyncTokenCredential
+    from azure.core.credentials import TokenCredential
 
 
 class ContainerServiceFleetMgmtClient:  # pylint: disable=client-accepts-api-version-keyword
     """Azure Kubernetes Fleet Manager api client.
 
     :ivar operations: Operations operations
-    :vartype operations:
-     azure.mgmt.containerservicefleet.v2023_08_15_preview.aio.operations.Operations
+    :vartype operations: azure.mgmt.containerservicefleet.v2023_10_15.operations.Operations
     :ivar fleets: FleetsOperations operations
-    :vartype fleets:
-     azure.mgmt.containerservicefleet.v2023_08_15_preview.aio.operations.FleetsOperations
+    :vartype fleets: azure.mgmt.containerservicefleet.v2023_10_15.operations.FleetsOperations
     :ivar fleet_members: FleetMembersOperations operations
     :vartype fleet_members:
-     azure.mgmt.containerservicefleet.v2023_08_15_preview.aio.operations.FleetMembersOperations
+     azure.mgmt.containerservicefleet.v2023_10_15.operations.FleetMembersOperations
     :ivar update_runs: UpdateRunsOperations operations
     :vartype update_runs:
-     azure.mgmt.containerservicefleet.v2023_08_15_preview.aio.operations.UpdateRunsOperations
+     azure.mgmt.containerservicefleet.v2023_10_15.operations.UpdateRunsOperations
     :ivar fleet_update_strategies: FleetUpdateStrategiesOperations operations
     :vartype fleet_update_strategies:
-     azure.mgmt.containerservicefleet.v2023_08_15_preview.aio.operations.FleetUpdateStrategiesOperations
+     azure.mgmt.containerservicefleet.v2023_10_15.operations.FleetUpdateStrategiesOperations
     :param credential: Credential needed for the client to connect to Azure. Required.
-    :type credential: ~azure.core.credentials_async.AsyncTokenCredential
+    :type credential: ~azure.core.credentials.TokenCredential
     :param subscription_id: The ID of the target subscription. Required.
     :type subscription_id: str
     :param base_url: Service URL. Default value is "https://management.azure.com".
     :type base_url: str
-    :keyword api_version: Api Version. Default value is "2023-08-15-preview". Note that overriding
-     this default value may result in unsupported behavior.
+    :keyword api_version: Api Version. Default value is "2023-10-15". Note that overriding this
+     default value may result in unsupported behavior.
     :paramtype api_version: str
     :keyword int polling_interval: Default waiting time between two polls for LRO operations if no
      Retry-After header is present.
@@ -61,7 +59,7 @@ class ContainerServiceFleetMgmtClient:  # pylint: disable=client-accepts-api-ver
 
     def __init__(
         self,
-        credential: "AsyncTokenCredential",
+        credential: "TokenCredential",
         subscription_id: str,
         base_url: str = "https://management.azure.com",
         **kwargs: Any
@@ -69,36 +67,32 @@ class ContainerServiceFleetMgmtClient:  # pylint: disable=client-accepts-api-ver
         self._config = ContainerServiceFleetMgmtClientConfiguration(
             credential=credential, subscription_id=subscription_id, **kwargs
         )
-        self._client: AsyncARMPipelineClient = AsyncARMPipelineClient(base_url=base_url, config=self._config, **kwargs)
+        self._client: ARMPipelineClient = ARMPipelineClient(base_url=base_url, config=self._config, **kwargs)
 
         client_models = {k: v for k, v in _models.__dict__.items() if isinstance(v, type)}
         self._serialize = Serializer(client_models)
         self._deserialize = Deserializer(client_models)
         self._serialize.client_side_validation = False
-        self.operations = Operations(
-            self._client, self._config, self._serialize, self._deserialize, "2023-08-15-preview"
-        )
-        self.fleets = FleetsOperations(
-            self._client, self._config, self._serialize, self._deserialize, "2023-08-15-preview"
-        )
+        self.operations = Operations(self._client, self._config, self._serialize, self._deserialize, "2023-10-15")
+        self.fleets = FleetsOperations(self._client, self._config, self._serialize, self._deserialize, "2023-10-15")
         self.fleet_members = FleetMembersOperations(
-            self._client, self._config, self._serialize, self._deserialize, "2023-08-15-preview"
+            self._client, self._config, self._serialize, self._deserialize, "2023-10-15"
         )
         self.update_runs = UpdateRunsOperations(
-            self._client, self._config, self._serialize, self._deserialize, "2023-08-15-preview"
+            self._client, self._config, self._serialize, self._deserialize, "2023-10-15"
         )
         self.fleet_update_strategies = FleetUpdateStrategiesOperations(
-            self._client, self._config, self._serialize, self._deserialize, "2023-08-15-preview"
+            self._client, self._config, self._serialize, self._deserialize, "2023-10-15"
         )
 
-    def _send_request(self, request: HttpRequest, **kwargs: Any) -> Awaitable[AsyncHttpResponse]:
+    def _send_request(self, request: HttpRequest, **kwargs: Any) -> HttpResponse:
         """Runs the network request through the client's chained policies.
 
         >>> from azure.core.rest import HttpRequest
         >>> request = HttpRequest("GET", "https://www.example.org/")
         <HttpRequest [GET], url: 'https://www.example.org/'>
-        >>> response = await client._send_request(request)
-        <AsyncHttpResponse: 200 OK>
+        >>> response = client._send_request(request)
+        <HttpResponse: 200 OK>
 
         For more information on this code flow, see https://aka.ms/azsdk/dpcodegen/python/send_request
 
@@ -106,19 +100,19 @@ class ContainerServiceFleetMgmtClient:  # pylint: disable=client-accepts-api-ver
         :type request: ~azure.core.rest.HttpRequest
         :keyword bool stream: Whether the response payload will be streamed. Defaults to False.
         :return: The response of your network call. Does not do error handling on your response.
-        :rtype: ~azure.core.rest.AsyncHttpResponse
+        :rtype: ~azure.core.rest.HttpResponse
         """
 
         request_copy = deepcopy(request)
         request_copy.url = self._client.format_url(request_copy.url)
         return self._client.send_request(request_copy, **kwargs)
 
-    async def close(self) -> None:
-        await self._client.close()
+    def close(self) -> None:
+        self._client.close()
 
-    async def __aenter__(self) -> "ContainerServiceFleetMgmtClient":
-        await self._client.__aenter__()
+    def __enter__(self) -> "ContainerServiceFleetMgmtClient":
+        self._client.__enter__()
         return self
 
-    async def __aexit__(self, *exc_details: Any) -> None:
-        await self._client.__aexit__(*exc_details)
+    def __exit__(self, *exc_details: Any) -> None:
+        self._client.__exit__(*exc_details)
