+++ src/fleet/azext_fleet/vendored_sdks/v2023_10_15/aio/_patch.py
