+++ src/fleet/azext_fleet/vendored_sdks/v2023_10_15/models/_container_service_fleet_mgmt_client_enums.py
